This PR will close https://github.com/apache/incubator-opendal/issues/1392.

By adding a new typed kv adapter, we can store metadata and bytes in rust containers like moka, dashmap, and avoid the extra cost of ser/de the bytes.

Before:

```rust
service_moka_read_full/4.00 KiB
                        time:   [3.1231 µs 3.1270 µs 3.1312 µs]
                        thrpt:  [1.2183 GiB/s 1.2199 GiB/s 1.2214 GiB/s]
service_moka_read_full/256 KiB
                        time:   [67.995 µs 68.043 µs 68.094 µs]
                        thrpt:  [3.5853 GiB/s 3.5881 GiB/s 3.5906 GiB/s]
service_moka_read_full/4.00 MiB
                        time:   [127.98 µs 128.19 µs 128.43 µs]
                        thrpt:  [30.415 GiB/s 30.473 GiB/s 30.522 GiB/s]
service_moka_read_full/16.0 MiB
                        time:   [731.54 µs 736.31 µs 741.73 µs]
                        thrpt:  [21.066 GiB/s 21.221 GiB/s 21.359 GiB/s]

service_moka_read_part/4.00 KiB
                        time:   [1.8117 µs 1.8238 µs 1.8349 µs]
                        thrpt:  [2.0790 GiB/s 2.0916 GiB/s 2.1056 GiB/s]
service_moka_read_part/256 KiB
                        time:   [24.402 µs 24.425 µs 24.449 µs]
                        thrpt:  [9.9856 GiB/s 9.9953 GiB/s 10.005 GiB/s]
service_moka_read_part/4.00 MiB
                        time:   [342.33 µs 342.99 µs 343.77 µs]
                        thrpt:  [11.363 GiB/s 11.389 GiB/s 11.411 GiB/s]
service_moka_read_part/16.0 MiB
                        time:   [6.1593 ms 6.1756 ms 6.1953 ms]
                        thrpt:  [2.5221 GiB/s 2.5301 GiB/s 2.5368 GiB/s]
```

After

```rust
service_moka_read_full/4.00 KiB
                        time:   [1.2101 µs 1.2125 µs 1.2151 µs]
                        thrpt:  [3.1394 GiB/s 3.1460 GiB/s 3.1525 GiB/s]
                 change:
                        time:   [-61.435% -61.358% -61.273%] (p = 0.00 < 0.05)
                        thrpt:  [+158.22% +158.79% +159.30%]
                        Performance has improved.
service_moka_read_full/256 KiB
                        time:   [5.4017 µs 5.4343 µs 5.4911 µs]
                        thrpt:  [44.462 GiB/s 44.926 GiB/s 45.197 GiB/s]
                 change:
                        time:   [-92.017% -91.914% -91.805%] (p = 0.00 < 0.05)
                        thrpt:  [+1120.2% +1136.7% +1152.6%]
                        Performance has improved.
service_moka_read_full/4.00 MiB
                        time:   [69.694 µs 71.709 µs 73.882 µs]
                        thrpt:  [52.872 GiB/s 54.474 GiB/s 56.049 GiB/s]
                 change:
                        time:   [-46.655% -45.354% -44.136%] (p = 0.00 < 0.05)
                        thrpt:  [+79.005% +82.995% +87.460%]
                        Performance has improved.
service_moka_read_full/16.0 MiB
                        time:   [225.82 µs 226.70 µs 227.88 µs]
                        thrpt:  [68.568 GiB/s 68.923 GiB/s 69.191 GiB/s]
                 change:
                        time:   [-69.957% -69.431% -68.914%] (p = 0.00 < 0.05)
                        thrpt:  [+221.69% +227.13% +232.86%]
                        Performance has improved.

service_moka_read_part/4.00 KiB
                        time:   [1.6380 µs 1.6405 µs 1.6434 µs]
                        thrpt:  [2.3212 GiB/s 2.3254 GiB/s 2.3289 GiB/s]
                 change:
                        time:   [-9.5756% -9.2733% -8.9819%] (p = 0.00 < 0.05)
                        thrpt:  [+9.8682% +10.221% +10.590%]
                        Performance has improved.
service_moka_read_part/256 KiB
                        time:   [6.0927 µs 6.1006 µs 6.1091 µs]
                        thrpt:  [39.964 GiB/s 40.019 GiB/s 40.071 GiB/s]
                 change:
                        time:   [-75.077% -75.015% -74.957%] (p = 0.00 < 0.05)
                        thrpt:  [+299.31% +300.24% +301.23%]
                        Performance has improved.
service_moka_read_part/4.00 MiB
                        time:   [95.818 µs 95.949 µs 96.095 µs]
                        thrpt:  [40.650 GiB/s 40.712 GiB/s 40.767 GiB/s]
                 change:
                        time:   [-72.177% -71.859% -71.610%] (p = 0.00 < 0.05)
                        thrpt:  [+252.23% +255.35% +259.41%]
                        Performance has improved.
service_moka_read_part/16.0 MiB
                        time:   [375.72 µs 378.70 µs 381.96 µs]
                        thrpt:  [40.908 GiB/s 41.260 GiB/s 41.586 GiB/s]
                 change:
                        time:   [-93.900% -93.826% -93.745%] (p = 0.00 < 0.05)
                        thrpt:  [+1498.6% +1519.8% +1539.4%]
                        Performance has improved.
```